# openApi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.0.1",
    "title": "Signal K Seamap Plugin API",
    "description": "Global sea charts based on OSM/Gebco/Emod with contour generation and bathymetry support",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [
    {
      "url": "/plugins/signalk-seamap-plugin",
      "description": "Plugin base path"
    }
  ],
  "paths": {
...
    "/soundings/{name}.json": {
      "get": {
        "tags": ["Soundings"],
        "summary": "Get soundings TileJSON metadata",
        "description": "Returns TileJSON metadata for spot soundings (depth measurements)",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "description": "DEM source name (mapterhorn, gebco, emod)",
            "schema": {
              "type": "string",
              "enum": ["mapterhorn", "gebco", "emod"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "TileJSON metadata with vector layer info",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tilejson": { "type": "string" },
                    "name": { "type": "string" },
                    "format": { "type": "string" },
                    "vector_layers": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "fields": {
                            "type": "object",
                            "properties": {
                              "depth": { "type": "string" }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Source not found"
          }
        }
      }
    },
    "/soundings/{name}/{z}/{x}/{y}.pbf": {
      "get": {
        "tags": ["Soundings"],
        "summary": "Get soundings tile",
        "description": "Generates and delivers spot soundings (depth point measurements) as vector tile (MVT/PBF) with jittered grid distribution",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "description": "DEM source name",
            "schema": {
              "type": "string",
              "enum": ["mapterhorn", "gebco", "emod"]
            }
          },
          {
            "name": "z",
            "in": "path",
            "required": true,
            "description": "Zoom level",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 14
            }
          },
          {
            "name": "x",
            "in": "path",
            "required": true,
            "description": "Tile X coordinate",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "y",
            "in": "path",
            "required": true,
            "description": "Tile Y coordinate",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Soundings vector tile with point features",
            "content": {
              "application/x-protobuf": {}
            }
          },
          "204": {
            "description": "No tile available"
          },
          "400": {
            "description": "Invalid tile coordinates"
          }
        }
      }
    },
...
  },
...
}